Restructured the project to better align with recommended practice.
